Move core change 70 to 74 in Ironwood

+++ openedx/core/djangoapps/content/course_overviews/models.py
@@ -223,36 +223,39 @@ class CourseOverview(TimeStampedModel):
         with store.bulk_operations(course_id):
             course = store.get_course(course_id)
             if isinstance(course, CourseDescriptor):
-                course_overview = cls._create_or_update(course)
-                try:
-                    with transaction.atomic():
-                        course_overview.save()
-                        # Remove and recreate all the course tabs
-                        CourseOverviewTab.objects.filter(course_overview=course_overview).delete()
-                        CourseOverviewTab.objects.bulk_create([
-                            CourseOverviewTab(tab_id=tab.tab_id, course_overview=course_overview)
-                            for tab in course.tabs
-                        ])
-                        # Remove and recreate course images
-                        CourseOverviewImageSet.objects.filter(course_overview=course_overview).delete()
-                        CourseOverviewImageSet.create(course_overview, course)
+                course_overview = CourseOverview.objects.filter(id=course.id).first()
+                if not course_overview:
+                    course_overview = cls._create_from_course(course)
+                    course_overview.save()
+                    try:
+                        with transaction.atomic():
+                            # Remove and recreate all the course tabs
+                            CourseOverviewTab.objects.filter(course_overview=course_overview).delete()
+                            CourseOverviewTab.objects.bulk_create([
+                                CourseOverviewTab(tab_id=tab.tab_id, course_overview=course_overview)
+                                for tab in course.tabs
+                            ])
+                            # Remove and recreate course images
+                            CourseOverviewImageSet.objects.filter(course_overview=course_overview).delete()
+                            CourseOverviewImageSet.create(course_overview, course)
 
-                except IntegrityError:
-                    # There is a rare race condition that will occur if
-                    # CourseOverview.get_from_id is called while a
-                    # another identical overview is already in the process
-                    # of being created.
-                    # One of the overviews will be saved normally, while the
-                    # other one will cause an IntegrityError because it tries
-                    # to save a duplicate.
-                    # (see: https://openedx.atlassian.net/browse/TNL-2854).
-                    pass
-                except Exception:
-                    log.exception(
-                        "CourseOverview for course %s failed!",
-                        course_id,
-                    )
-                    raise
+                    except IntegrityError:
+                        # There is a rare race condition that will occur if
+                        # CourseOverview.get_from_id is called while a
+                        # another identical overview is already in the process
+                        # of being created.
+                        # One of the overviews will be saved normally, while the
+                        # other one will cause an IntegrityError because it tries
+                        # to save a duplicate.
+                        # (see: https://openedx.atlassian.net/browse/TNL-2854).
+                        log.error("IntegrityError error for '%s'" % course_id)
+                        pass
+                    except Exception:
+                        log.exception(
+                            "CourseOverview for course %s failed!",
+                            course_id,
+                        )
+                        raise
 
                 return course_overview
             elif course is not None:
@@ -558,7 +561,7 @@ class CourseOverview(TimeStampedModel):
         log.info('Finished generating course overviews.')
 
     @classmethod
-    def get_all_courses(cls, orgs=None, filter_=None):
+    def get_all_courses(cls, orgs=None, filter_=None, exclude_=None):
         """
         Returns all CourseOverview objects in the database.
 
@@ -566,6 +569,7 @@ class CourseOverview(TimeStampedModel):
             orgs (list[string]): Optional parameter that allows case-insensitive
                 filtering by organization.
             filter_ (dict): Optional parameter that allows custom filtering.
+            exclude_ (dict): Optional parameter that allows custom excludes.
         """
         # Note: If a newly created course is not returned in this QueryList,
         # make sure the "publish" signal was emitted when the course was
@@ -580,6 +584,9 @@ class CourseOverview(TimeStampedModel):
 
         if filter_:
             course_overviews = course_overviews.filter(**filter_)
+
+        if exclude_:
+            course_overviews = course_overviews.exclude(**exclude_)
 
         return course_overviews
